validate rates
A PGM may have an invalid specification if two
incoming edges fire in a way that causes one edge
to overflow or underflow. This patch adds PGM
validation. In the future, it would be nice to
patch-up an invalid graph to make it valid. This
patch-up would entail the addition of dummy tokens
to edges with deficient production values.