Use posix mutex and cond var to reduce ctxs.
This patch allows libpgm to be compiled with
posix and condition variables that are used
to optimize token transmission. With this
feature enabled, consumer nodes block on a
condition variable instead of on select(fifos)
within a loop. The consumer blocked on a
condition variable is ONLY signalled when its
token constraints have been satisfied. Thus,
context switches are reduced. Furthermore,
this framework also makes PGM consumer thresholds
trivial to implement, so that feature is also
added by this patch.