Boost producers, not consumers.
Overhead accounting is easier (and more
accurate) if producers run at a boosted
priority when they are sending out tokens.
Prior this patch, only consumers waiting
for inbound tokens ran at a boosted priority
(which was incorrect). This was incorrect
because a boosted consumer can prevent its
unboosted producer from signalling/waking
a higher-priority sibling consumer,
resulting in a priority inversion. This could
have been addressed by also boosting the
priority of the producer (and made to have
a higher prio that any of its boosted
consumers), but that also introduces priority
inversions. Instead, it's best to only boost
the producer and shift the deadline of the
consumer while it is waiting for tokens, as
per Liu's release analysis.